Conditionally apply 53ns offset

+++ hedera-mirror-importer/src/main/resources/db/migration/v1/V1.65.7__reconciliation_job.sql
@@ -1,7 +1,7 @@
 create table if not exists reconciliation_job
 (
     consensus_timestamp bigint      not null,
-    count               int         not null,
+    count               bigint      not null,
     error               text        not null default '',
     timestamp_end       timestamptz null,
     timestamp_start     timestamptz not null,
