+++ hedera-mirror-importer/src/main/resources/db/migration/v2/V2.0.0__create_tables.sql
@@ -437,7 +437,7 @@ comment on table prng is 'Pseudorandom number generator';
 create table if not exists reconciliation_job
 (
     consensus_timestamp bigint      not null,
-    count               int         not null,
+    count               bigint      not null,
     error               text        not null default '',
     timestamp_end       timestamptz null,
     timestamp_start     timestamptz not null,
